Try monitor restart

+++ userspace/dragent/monitor.cpp
@@ -72,6 +72,7 @@ int monitor::run()
 		auto child_pid = fork();
 		if(child_pid < 0)
 		{
+			delete_pid_file(m_pidfile);
 			exit(EXIT_FAILURE);
 		}
 		else if(child_pid == 0)
@@ -102,13 +103,38 @@ int monitor::run()
 			}
 			else if(waited_pid > 0)
 			{
-				if(process.is_main() && WIFEXITED(status) && WEXITSTATUS(status) == 0)
+				if(process.is_main() && WIFEXITED(status))
 				{
-					//
-					// Process terminated cleanly
-					//
-					delete_pid_file(m_pidfile);
-					exit(EXIT_SUCCESS);
+					if(WEXITSTATUS(status) == 0)
+					{
+						//
+						// Process terminated cleanly
+						//
+						delete_pid_file(m_pidfile);
+						exit(EXIT_SUCCESS);
+					}
+					else if(WEXITSTATUS(status) == CONFIG_UPDATE_EXIT_CODE)
+					{
+						for(const auto& process : m_processes)
+						{
+							//
+							// Send TERM to all others
+							if(!process.is_main())
+							{
+								if(kill(process.pid(), SIGKILL) != 0)
+								{
+									delete_pid_file(m_pidfile);
+									exit(EXIT_FAILURE);
+								}
+								// FIXME: locking wait here
+								waitpid(process.pid(), NULL, 0);
+							}
+						}
+						execl("/opt/draios/bin/dragent", "dragent", NULL);
+
+						delete_pid_file(m_pidfile);
+						exit(EXIT_FAILURE);
+					}
 				}
 
 				if(!process.is_main())
@@ -139,6 +165,7 @@ int monitor::run()
 				auto child_pid = fork();
 				if(child_pid < 0)
 				{
+					delete_pid_file(m_pidfile);
 					exit(EXIT_FAILURE);
 				}
 				else if(child_pid == 0)
@@ -160,14 +187,17 @@ int monitor::run()
 		// Signal received, forward it to the child and
 		// wait for it to terminate
 		//
-		if(kill(process.pid(), g_signal_received) != 0)
+		if(process.pid() > 0)
 		{
-			delete_pid_file(m_pidfile);
-			exit(EXIT_FAILURE);
-		}
-		if(process.is_main())
-		{
-			waitpid(process.pid(), NULL, 0);
+			if(kill(process.pid(), g_signal_received) != 0)
+			{
+				delete_pid_file(m_pidfile);
+				exit(EXIT_FAILURE);
+			}
+			if(process.is_main())
+			{
+				waitpid(process.pid(), NULL, 0);
+			}
 		}
 	}
